+++ front/src/environments/environment.ts
@@ -1,3 +1,5 @@
 export const environment = {
-  googleClientId: 'GOOGLE_CLIENT_ID',
+  production: false,
+  googleClientId:
+    '464908390325-c5a91uln62d88duvrd1jemue65isr1hv.apps.googleusercontent.com',
 };